+++ CHANGELOG.md
@@ -46,6 +46,7 @@
     - `hasParticipantFlagKey` new expression, that will check if a participant has a flag attribute with a specific key, where the value is not checked.
     - `getParticipantFlagValue` new expression, that will retrieve the value of a participant flag with a given key.
     - `hasResponseKey` new expression, that will return true if the question has a response that contains a key at the specified path (e.g.: T0.Q1, rg.scg.1.b)
+    - `getSelectedKeys` to retrieve the list of keys from a response slot, encoded as a string, separated by semicolons `;`
     - `hasResponseKeyWithValue` new expression, that will return true if the question has a response that contains a key at the specified path with the given value (e.g.: T0.Q1, rg.scg.1.b, value)
     - New actions for working with reports: `INIT_REPORT, UPDATE_REPORT_DATA, REMOVE_REPORT_DATA, CANCEL_REPORT`. At the start o the study event, a map of reports is initalised (empty), and during the event, actions can create one or more report entries. Reports that are in this map at the end of the event will be saved to the database.
     - implemented new expressions to handle merge event (when two participant states should be merged - convert temporary participant when participant already exists)